머지전 확인

+++ Project/src/main/java/com/mvc/grade/Grade.java
@@ -28,5 +28,5 @@ public class Grade {
 	private Long userCode;
 	
 	@Column(nullable = true)
-	private Long grade;
+	private Double grade;
 }
